requirements/common.txt
Flask==0.12
Flask-Admin==1.5.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-SQLAlchemy==2.1
Flask-Migrate==2.0.3
flask-nav==0.6
Flask-Script==2.0.5
Flask-WTF==0.14.2
python-dateutil==2.6.1
marshmallow==2.13
marshmallow-sqlalchemy==0.13
Jinja2==2.9.5
SQLAlchemy==1.1.5
WTForms==2.1
Werkzeug==0.11.15
alembic==0.8.10
click==6.7
python-editor==1.0.3